wasm: fix symbol emitting for relocatables
This fixes the following points to ensure the symbols emitted are
serialized correctly:
- Unmark __wasm_call_ctors for relocatables as it's a synthetic symbol
and should not be part of the relocatable object file.
- Update the index of symbols pointing to debug and data sections.
The previous index were the internal representation, rather than what
would be emitted into the relocatable object file. Now we update the
internal representation lazily during writing.