wasm: emit correct relocation index for types
Type index relocations have the `index` field point directly to the index
of the type within the list of types, rather than a symbol. This means
that the index no longer corresponds to the right index after we merge
all type indexes. So when we emit the relocation, ensure we retrieve the
new type index on-demand as we emit the relocation.